Possibly controversial readme update (#166)
* Possibly controversial statements on design patterns

This is a rough draft and hits most of the concepts that seem to confuse people who use the EDS but haven't developed in it. 

I think it's important to get the messaging on this right so people can understand what the unwritten rules of the EDS design are. The first step is for us to all agree on them, this is a very rough draft and seems likely to create disagreements. My goal here was mostly to record things as they are now, without providing any "we'd like to do X eventually" commentary.

* Possibly controversial statements on design patterns

This is a rough draft and hits most of the concepts that seem to confuse people who use the EDS but haven't developed in it.

I think it's important to get the messaging on this right so people can understand what the unwritten rules of the EDS design are. The first step is for us to all agree on them, this is a very rough draft and seems likely to create disagreements. My goal here was mostly to record things as they are now, without providing any "we'd like to do X eventually" commentary.

* Respond to github comments

* v1.1.20

* v1.1.21

* v1.1.26